Fix database preloading to only preload when database is empty
