+++ questionbank/src/main/java/com/example/questionbank/QuestionbankApplication.java
@@ -1,54 +1,13 @@
 package com.example.questionbank;
 
-import com.example.questionbank.model.Question;
-import com.example.questionbank.repository.QuestionRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.util.Arrays;
 
 @SpringBootApplication
 
 public class QuestionbankApplication {
-
-    private static final Logger log = LoggerFactory.getLogger(QuestionbankApplication.class);
-
     public static void main(String... args) {
         SpringApplication.run(QuestionbankApplication.class, args);
     }
 
-    @Bean
-    CommandLineRunner initDatabase(QuestionRepository repository) {
-
-        return args -> {
-            log.info("Preloading " + repository.save(new Question("Reverse a String",
-                    "Write a function that reverses a string. The input string is given as an array of characters s. \n" +
-                            "\n" +
-                            "You must do this by modifying the input array in-place with O(1) extra memory. \n" +
-                            "\n" +
-                            "Example 1:\n" +
-                            "\n" +
-                            "Input: s = [\"h\",\"e\",\"l\",\"l\",\"o\"]\n" +
-                            "Output: [\"o\",\"l\",\"l\",\"e\",\"h\"]\n" +
-                            "\n" +
-                            "Example 2:\n" +
-                            "Input: s = [\"H\",\"a\",\"n\",\"n\",\"a\",\"h\"]\n" +
-                            "Output: [\"h\",\"a\",\"n\",\"n\",\"a\",\"H\"]\n" +
-                            "\n" +
-                            "Constraints:\n" +
-                            "\n" +
-                            "1 <= s.length <= 105\n" +
-                            "s[i] is a printable ascii character.\n",
-                    Arrays.asList("Strings", "Algorithms"), // Multiple categories
-                    "Easy")));
-            log.info("Preloading " + repository.save(new Question("Linked List Cycle Detection",
-                    "Implement a function to detect if a linked list contains a cycle.",
-                    Arrays.asList("Data Structures", "Algorithms"), // Multiple categories
-                    "Easy")));
-        };
-    }
 }
